Fix some missing headers and compilation issues on linux
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@6262 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/bignum/polynomial.cc
@@ -296,7 +296,7 @@ string Polynomial::ToCode(const string &type, const Polynomial &p) {
                 "\n"
                 "  return ");
   {
-    for (int sidx = 0; sidx < p.sum.size(); sidx++) {
+    for (int sidx = 0; sidx < (int)p.sum.size(); sidx++) {
       if (sidx != 0)
         StringAppendF(&code, " + ");
       StringAppendF(&code, "%s", SummandVar(sidx).c_str());
